octeontx: advertise local LBK index of loop index
ODP will need to know which of LBK/virt interfaces is
configured as loop. Pass this index in SSO_GETDOMAINCFG response.

Reviewed-on: https://sj1git1.cavium.com/13076
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>

+++ drivers/net/ethernet/cavium/octeontx-83xx/pkopf_main.c
@@ -295,7 +295,7 @@ static struct intr_hand intr[] = {
 static void identify(struct pkopf_vf *vf, u16 domain_id,
 		     u16 subdomain_id)
 {
-	u64 reg = (((u64)subdomain_id << 16) | (domain_id)) << 7;
+	u64 reg = ((subdomain_id) << 16 | (domain_id)) << 7;
 
 	writeq_relaxed(reg, vf->domain.reg_base + PKO_VF_DQ_FC_CONFIG);
 }
